add formula to calculate the rrtd

+++ SRCDIR/temp/temp_measure.h
@@ -4,7 +4,7 @@
 int TempMeasInit(void);
 int TempMeasStart(void);
 int TempMeasStop(void);
-UINT32 TempMeasCalibration(void);
+float TempMeasCalibration(void);
 void TempMeasDump(void);
 
 #endif
